Add Neomake support.
Own maker (build) to work with cmake projects.
Search closest build folder, and append it to maker.
If using clang, NeomakeBuildErrorFormatClang() to parse output properly.
If gcc, xxxGCC().

If you need to modify the buildFolder, NeomakeBuildUpdate('to/new/path')
will update neomake.

Also a NeomakeToggle() is provided to build or not on save.
One shot compilation can be done with :call NeomakeBuild().

Common pipeline:
:call NeomakeBuildPrepare()
:call NeomakeBuildErrorFormatClang()
:call ToggleNeomakeBuildOnSave()

or easier:
:call NeomakeBuildPrepareClang()