Replace vim-airline for lightline
Mainly because the massive delay when using Ctrl-^,
and any other vim functionality on buffers.
airline seemed to iterate over every open buffer, and making the
function slow.
Lightline is snap fast.

Cons: Colors in lightline are not tweakable.

I also added integration with AsyncRun, the status of it is now
showed in the right of the tabline (buffers are shown in the left)